Mappping Models and swagger-2 ui changes

+++ src/main/java/com/roni/qeats/configs/AppConstants.java
@@ -6,6 +6,6 @@ public class AppConstants {
 	public static final String PAGE_SIZE = "10";
 	public static final String SOTRT_BY = "postId";
 	public static final String SORT_DIR = "asc";
-	public static final Integer NORMAL_USER = 502;
-	public static final Integer ADMIN_USER = 501;
+	public static final Long NORMAL_USER = 502L;
+	public static final Long ADMIN_USER = 501L;
 }
